Add files via upload
C++ code to create 4 functions to calculate addition, subtraction, multiplication, and division from user-inputted values for X and Y. Results will be outputted to the command prompt. If/else loops were used to control specific use cases in multiplication and division. A main function is created to take user inputs for X and Y.